mostly working decoder
