+++ Source/DataTypes/Object.swift
@@ -78,31 +78,32 @@ enum Object: Equatable, Codable {
     }
 
     mutating func set(value: Object, at keyPath: [String]) {
+        if keyPath.count == 0 {
+            self = value
+            return
+        }
         var keyPath = keyPath
         if case .map(var map) = self {
             let key = keyPath.removeFirst()
-            if keyPath.count == 0 {
-                map[key] = value
-            } else {
-                var valueAtKey = map[key] ?? .map([:])
+            if var valueAtKey = map[key] {
                 valueAtKey.set(value: value, at: keyPath)
                 map[key] = valueAtKey
+            } else {
+                map[key] = value
             }
             self = .map(map)
         } else if case .list(var list) = self {
-            let string = keyPath.removeFirst()
-            let key = Int(string)!
-            if keyPath.count == 0 {
-                list[key] = value
+            let key = Int(keyPath.removeFirst())!
+            if list.count <= key {
+                list.append(value)
             } else {
-                var valueAtKey = list[key] //?? .list([])
+                var valueAtKey = list[key]
                 valueAtKey.set(value: value, at: keyPath)
                 list[key] = valueAtKey
             }
             self = .list(list)
         }
     }
-
 }
