+++ Source/Proxy/Proxy+Collection.swift
@@ -76,8 +76,10 @@ extension Proxy: RangeReplaceableCollection where Wrapped: RangeReplaceableColle
         let newElements = proxyElements.map { $0.get() }
         let start = subrange.relative(to: self).startIndex
         let deleteCount = subrange.relative(to: self).endIndex - subrange.relative(to: self).startIndex
-        let encoded = try!TypeToObject().map(Array(newElements))
-        context.splice(path: path, start: start, deletions: deleteCount, insertions: encoded)
+        guard case .list(let list) = try! objectEncoder.encode(Array(newElements)) else {
+            fatalError()
+        }
+        context.splice(path: path, start: start, deletions: deleteCount, insertions: Array(list))
     }
 
     public func reserveCapacity(_ n: Int) {}
@@ -89,8 +91,10 @@ extension Proxy where Wrapped: RangeReplaceableCollection, Wrapped.Index == Int,
     public func replaceSubrange<C, R>(_ subrange: R, with newElements: C) where C : Collection, R : RangeExpression, C.Element == Wrapped.Element, Index == R.Bound {
         let start = subrange.relative(to: self).startIndex
         let deleteCount = subrange.relative(to: self).endIndex - subrange.relative(to: self).startIndex
-        let encoded = try! TypeToObject().map(Array(newElements))
-        context.splice(path: path, start: start, deletions: deleteCount, insertions: encoded)
+        guard case .list(let list) = try! objectEncoder.encode(Array(newElements)) else {
+            fatalError()
+        }
+        context.splice(path: path, start: start, deletions: deleteCount, insertions: Array(list))
     }
 
     public func append(_ newElement: __owned Wrapped.Element) {
