+++ Source/Proxy/Proxy.swift
@@ -40,7 +40,7 @@ public final class Proxy<Wrapped: Codable> {
     private let valueResolver: () -> Wrapped?
 
     lazy var objectToType = ObjectToTypeTransformer()
-    lazy var typeToObject = ObjectToTypeTransformer()
+    let objectEncoder = ObjectEncoder()
 
     public func get() -> Wrapped {
         return valueResolver()!
@@ -101,8 +101,7 @@ public final class Proxy<Wrapped: Codable> {
     }
 
     public func set(_ newValue: Wrapped) {
-        let mapper = TypeToObject()
-        let object = try! mapper.map(newValue)
+        let object = try! objectEncoder.encode(newValue)
 
         set(newValue: object)
     }
